+++ covobeta/src/components/TestPage/index.js
@@ -8,7 +8,8 @@ import React, { Component } from "react";
 import "./TestPage.css";
 
 //Temporary or unclassified imports
-import AddTrip from "../AddTrip";
+//import AddTrip from "../AddTrip";
+import CoVoTripsFromContext from "./Tests/CoVoTripsFromContext.js";
 
 //Descriptions of props, what it does, and what it affects
 //Beginning of implementation
@@ -16,7 +17,8 @@ class TestPage extends Component {
   render() {
     return (
       <div className="test-page">
-        <AddTrip />
+        {/*<AddTrip />*/}
+        <CoVoTripsFromContext />
       </div>
     );
   }