+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## :arrow_down: v0.0.2 Starting work on adding trip
 
 * Now will work on pushing to firebase.
-  * Worked a little on javascript firestore/firebase in db, it loads.
+  * worked a little on javascript firestore/firebase in db, it loads.
 * Working on AddTrip which should upload trip to cloud from Props
   * will work with [redux-firestore](https://github.com/CoVoCre/redux-firestore) and [react-redux-firebase](https://github.com/CoVoCre/react-redux-firebase), must add these to redux store creation process
   * added TestPage and also to router in order to render components in developpment there